requirements.txt
asgiref==3.10.0
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
Brotli==1.1.0
cbor2==5.7.1
certifi==2025.10.5
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
cryptography==46.0.3
cssselect2==0.8.0
daphne==4.2.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-extra-fields==3.7.0
et_xmlfile==2.0.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.11
ImageIO==2.37.2
Incremental==24.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.45.1
mpmath==1.3.0
msgpack==1.1.2
mysqlclient==2.2.7
networkx==3.5
numba==0.62.1
numpy==2.2.6
onnxruntime==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
protobuf==6.33.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
PyMatting==1.1.14
pyOpenSSL==25.3.0
pyphen==0.17.2
pyreadline3==3.5.4
pytesseract==0.3.13
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
qrcode==8.2
redis==7.1.0
referencing==0.37.0
rembg==2.0.67
reportlab==4.4.4
requests==2.32.5
rpds-py==0.28.0
scikit-image==0.25.2
scipy==1.16.3
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.10.16
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
zope.interface==8.1.1
zopfli==0.2.3.post1
